CI and Build Improvements (#687)
    Remove VS2015 build from CI.
    Move Windows builds (except VS2017) to GitHub Actions.
    Run Windows builds and tests with both MSVC and CLangCL on VS2022.
    Build one configuration with VS2019 to ensure it still works.
    Build Windows release packages with VS2022 and CLangCL.
    Turn on warnings as errors. Fix bug that prevented /W4 for CLangCL. Fix all warnings.
    Update the included zstd.c and regen golden images for affected tests.
    Stop using deprecated code in GitHub Actions.
    Update Linux CI runner to Ubuntu Jammy
    Update Vulkan SDK used in Linux and Windows CI to 1.3.243.
    Fix issue with Doxygen builds that caused intermittent failures.

+++ lib/texture2.c
@@ -2321,7 +2321,7 @@ ktxTexture2_inflateZstdInt(ktxTexture2* This, ktx_uint8_t* pDeflatedData,
     DECLARE_PROTECTED(ktxTexture);
     ktx_uint32_t levelIndexByteLength =
                             This->numLevels * sizeof(ktxLevelIndexEntry);
-    uint32_t levelOffset = 0;
+    uint64_t levelOffset = 0;
     ktxLevelIndexEntry* cindex = This->_private->_levelIndex;
     ktxLevelIndexEntry* nindex;
     ktx_uint32_t uncompressedLevelAlignment;
